[metasrv] refactor: move test cases into metasrv/raft/state_machine/testing.rs
