+++ metasrv/src/meta_service/meta_service_impl_test.rs
@@ -41,7 +41,7 @@ async fn test_meta_server_add_file() -> anyhow::Result<()> {
 
     let mut client = MetaServiceClient::connect(format!("http://{}", addr)).await?;
 
-    let cases = crate::meta_service::raftmeta_test::cases_add_file();
+    let cases = crate::raft::state_machine::testing::cases_add_file();
 
     for (name, txid, k, v, want_prev, want_rst) in cases.iter() {
         let req = LogEntry {
@@ -82,7 +82,7 @@ async fn test_meta_server_set_file() -> anyhow::Result<()> {
 
     let mut client = MetaServiceClient::connect(format!("http://{}", addr)).await?;
 
-    let cases = crate::meta_service::raftmeta_test::cases_set_file();
+    let cases = crate::raft::state_machine::testing::cases_set_file();
 
     for (name, txid, k, v, want_prev, want_rst) in cases.iter() {
         let req = LogEntry {
@@ -221,7 +221,7 @@ async fn test_meta_server_incr_seq() -> anyhow::Result<()> {
 
     let mut client = MetaServiceClient::connect(format!("http://{}", addr)).await?;
 
-    let cases = crate::meta_service::raftmeta_test::cases_incr_seq();
+    let cases = crate::raft::state_machine::testing::cases_incr_seq();
 
     for (name, txid, k, want) in cases.iter() {
         let req = LogEntry {
